[Delievrs #160545778] move routes to api file in models
